Move bin/hex conversion functions out of dfu commands
