+++ src/arguments.c
@@ -28,6 +28,7 @@
 #include "dfu-device.h"
 #include "config.h"
 #include "arguments.h"
+#include "util.h"
 
 // Modes used to display the list of targets.
 #define LIST_STD        0
@@ -52,6 +53,10 @@
 #define BL_SPECIFIC     3   /* Any value greater than this is a specific start address */
 
 extern int debug;       /* defined in main.c */
+
+#define ARGUMENTS_DEBUG_THRESHOLD 100
+
+#define DEBUG(...)  dfu_debug( __FILE__, __FUNCTION__, __LINE__, ARGUMENTS_DEBUG_THRESHOLD, __VA_ARGS__ )
 
 struct option_mapping_structure {
     const char *name;
@@ -1056,6 +1061,134 @@ static void print_args( struct programmer_arguments *args )
     fflush( stdout );
 }
 
+static int32_t execute_hex2bin( struct programmer_arguments *args ) {
+    int32_t  retval = -1;
+    uint32_t  i;
+    intel_buffer_out_t bout;
+    size_t   memory_size;
+    size_t   page_size;
+    uint32_t target_offset = 0;
+
+    memory_size = args->memory_address_top + 1;
+    page_size = args->flash_page_size;
+
+    // ----------------- CONVERT HEX FILE TO BINARY -------------------------
+    if( 0 != intel_init_buffer_out(&bout, memory_size, page_size) ) {
+        DEBUG("ERROR initializing a buffer.\n");
+        goto error;
+    }
+
+    if( 0!= intel_hex_to_buffer( args->com_convert_data.file, &bout,
+                target_offset, args->quiet ) ) {
+        DEBUG( "Something went wrong with creating the memory image.\n" );
+        goto error;
+    }
+
+    if( !args->quiet )
+        fprintf( stderr, "Dumping 0x%X bytes from address offset 0x%X.\n",
+                bout.info.data_end + 1, target_offset );
+    for( i = 0; i <= bout.info.data_end; i++ ) {
+        fprintf( stdout, "%c", bout.data[i] <= 0xFF ? bout.data[i] & 0xFF : 0xFF);
+    }
+
+    fflush( stdout );
+
+    retval = 0;
+
+error:
+    if( NULL != bout.data ) {
+        free( bout.data );
+        bout.data = NULL;
+    }
+
+    return retval;
+}
+
+static int32_t execute_bin2hex( struct programmer_arguments *args ) {
+    int32_t retval = -1;        // return value for this fcn
+    intel_buffer_in_t buin;     // buffer in for storing read mem
+    enum atmel_memory_unit_enum mem_segment = args->com_convert_data.segment;
+    size_t mem_size = 0;
+    size_t page_size = 0;
+    uint32_t target_offset = 0; // address offset on the target device
+        // NOTE: target_offset may not be set appropriately for device
+        // classes other than ADC_AVR32
+    FILE *fp = NULL;
+    char *filename = args->com_convert_data.file;
+
+    if( ADC_AVR32 == args->device_type ) {
+        target_offset = 0x80000000;
+    }
+
+    switch( mem_segment ) {
+        case mem_flash:
+            mem_size = args->memory_address_top + 1;
+            page_size = args->flash_page_size;
+            break;
+        case mem_eeprom:
+            mem_size = args->eeprom_memory_size;
+            page_size = args->eeprom_page_size;
+            break;
+        case mem_user:
+            mem_size = args->flash_page_size;
+            page_size = args->flash_page_size;
+            target_offset = 0x80800000;
+            break;
+        default:
+            fprintf( stderr, "Dump not currently supported for this memory.\n" );
+            goto error;
+    }
+
+    if( 0 != intel_init_buffer_in(&buin, mem_size, page_size) ) {
+        DEBUG("ERROR initializing a buffer.\n");
+        goto error;
+    }
+
+    if( mem_segment == mem_flash ) {
+        buin.info.data_start = args->flash_address_bottom;
+        buin.info.data_end = args->flash_address_top;
+    }
+
+    if( NULL == filename ) {
+        if( !args->quiet ) fprintf( stderr, "Invalid filename.\n" );
+        retval = -2;
+        goto error;
+    }
+
+    if( 0 == strcmp("STDIN", filename) ) {
+        fp = stdin;
+    } else {
+        fp = fopen( filename, "r" );
+        if( NULL == fp ) {
+            if( !args->quiet ) fprintf( stderr, "Error opening %s\n", filename );
+            retval = -3;
+            goto error;
+        }
+    }
+
+    buin.info.data_end = fread(buin.data, 1, buin.info.total_size, fp);
+    if( buin.info.data_end == 0 ) {
+        if( !args->quiet ) fprintf( stderr, "ERROR: no bytes read\n" );
+        retval = -4;
+        goto error;
+    }
+
+    if( !args->quiet )
+        fprintf( stderr, "Read 0x%X bytes, making hex with address offset 0x%X.\n",
+                buin.info.data_end + 1, target_offset );
+
+    retval = intel_hex_from_buffer( &buin, args->com_convert_data.force, target_offset );
+
+error:
+    if( NULL != buin.data ) {
+        free( buin.data );
+        buin.data = NULL;
+    }
+
+    return retval;
+}
+
+
 int32_t parse_arguments( struct programmer_arguments *args,
                          const size_t argc,
                          char **argv )
@@ -1182,6 +1315,12 @@ int32_t parse_arguments( struct programmer_arguments *args,
             goto done;
         }
         args->com_convert_data.file[0] = args->com_convert_data.original_first_char;
+
+        if ( com_bin2hex == args->command ) {
+            return execute_bin2hex( args );
+        } else {
+            return execute_hex2bin( args );
+        }
     }
 
 done:
